requirements.txt
numpy==1.26.3
openai==1.14.0
flask==2.3.3
werkzeug==2.3.7
python-dotenv==1.0.0